Move person creators to display
